backend/DjangoServer/requirements.txt
djangorestframework
markdown
django-filter
django-cors-headers
Pillow